+++ packages/leancode_contracts_generator/example/ExampleContracts/Users/EditUser.cs
@@ -1,13 +1,13 @@
 using System;
 using System.Collections.Generic;
-using LeanCode.CQRS;
-using LeanCode.CQRS.Security;
+using LeanCode.Contracts;
+using LeanCode.Contracts.Security;
 
 namespace LeanCode.ContractsGeneratorV2.ExampleContracts.Users
 {
     [AuthorizeWhenHasAnyOf(Auth.Roles.Admin)]
     [Security.AuthorizeWhenHasSomethingAccess]
-    public class EditUser : IRemoteCommand, Security.ISomethingRelated
+    public class EditUser : ICommand, Security.ISomethingRelated
     {
         public Guid UserId { get; set; }
         public Guid SomethingId { get; set; }